+++ R/trip-deprecated.R
@@ -8,8 +8,7 @@
 #' @param x see \code{\link{cut.trip}}
 #' @param dates see \code{\link{cut.trip}}
 #' @param from trip object
-#' @param crs CRS object, or PROJ.4 string accepted by \code{\link[sp]{CRS}}
-#' @param \dots Further arguments to \code{\link[rgdal]{spTransform}}
+#' @param \dots Further arguments 
 #' @seealso
 #' 
 #' \code{\link{cut.trip}}, \code{\link{as.Other}}
@@ -48,13 +47,3 @@ as.trip.SpatialLinesDataFrame <- function(from) {
   as(from, "SpatialLinesDataFrame")
 }
 
-## removed depends sp, suggests rgdal deprecate this, replace with
-## spTransform method below
-
-##' @rdname trip.split.exact
-##' @export
-tripTransform <- function(x, crs, ...) {
-  .Deprecated("spTransform")
-  if (! inherits(crs, "CRS")) crs <- CRS(crs)
-  spTransform(x, crs, ...)
-}