+++ NEWS.md
@@ -1,5 +1,7 @@
 # trip (development version)
 
+* Migrate to reproj to avoid rgdal package. Ancient function tripTransform has been removed completely, and spTransform methods now defunct pointing to `reproj()`. Ultimately we want a pure PROJ package or equivalent for these basic transformations, and proj4 is the best for now.  
+
 * `readDiag()` and `readArgos()` now read all records from a single text vector after reading
 all raw lines. This means they can be used to read data from mixed text streams.